requirements.txt
python-docx==0.2.4
openpyxl==3.1.2
pandas==1.4.3
Pillow==10.0.0
pywin32==304
selenium==4.10.0
